# Improve policy import performance