_Story time!_

When CIDR policy support was initially introduced, we added dedicated           
code to parse and merge incoming CIDR policy rules into an internal             
`CIDRPolicy` format, and used that implementation to plumb CIDR-specific        
maps in the eBPF datapath.                                                      
                                                                                
However, soon after it became clear that having dedicated CIDR policy           
code in the datapath would be problematic when attempting to support a          
variety of more complicated policies:
* Label-based policy would  be handled separately and could fall
  out-of-sync with the CIDR-based policy logic.
* If we wished to support AND operations for L3 and L4 statements
  ("allow to 1.1.1.1 on port 80"), we would have to expand the map keys
  and lookup code for each of these paths to handle these cases separately.
* Even if the code maintenance issues were not a concern, then verifier
  complexity could cause problems over time due to the extra instructions
  necessary for implementing policy in this way.
* Finally, the notion of executing additional instructions on a per-packet
  basis to do another slightly different type of policy lookup in the datapath
  did not sit well from the perspective of performance and improving latency.
                                                                                
PR #3835 resolved the above issues by reworking the policy calculation          
logic for CIDR policies to first pre-process the CIDR policy rules into         
`LabelSelector`s, which are already used for selecting pods internally in         
the main policy engine. This allowed CIDR policy to be generated from           
the same internal structure as regular label-based policy, and served as        
a basis for centralizing the datapath enforcement around a single               
policymap per endpoint.                                                         
                                                                                
That said, there was one corner of policy which still relied on the old         
`CIDRPolicy` calculation code: Handling IPCache lookups for CIDRs. On             
older kernels which lack support for eBPF LPM map types (eg 4.9), when          
generating per-endpoint datapath code, Cilium would look at all of the          
actively-used CIDR rules and determine all of the unique prefixes, and          
then use this to construct a makeshift LPM map out of hash maps (with           
each unique prefix requiring a dedicated unrolled loop iteration to             
perform the lookup).                                                            
                                                                                
Commit 3096bda65030 ("daemon: Track prefix lengths in use in policy")           
already started tracking these on a daemon-wide basis a long time ago,          
though this doesn't minimize the complexity for each endpoint; if just          
one too many CIDR policies would be added, then it could cause                  
complexity issues for all endpoints rather than only those affected by          
the policy. Between this and shifted focus on improving the policy              
engine elsewhere, we ended up just leaving this code as-is until now.           
                                                                                
Over time, users have migrated on to kernels with a real LPM                    
implementation. Furthermore, other policy types like label-based                
policies and FQDN policy have allowed CIDR policies to take a back seat,        
meaning that users have typically not pushed the limits of CIDR policies        
particularly on those older kernels. By now, Linux 4.9 is getting a bit         
long in the tooth, and there are many reasons for users to run newer            
kernels, especially for better eBPF support. So, I think it's finally           
time to retire this CIDR-specific policy generation code. As a bonus,           
we'll even get a performance bump in policy processing times.                   
                                                                                
Best result of five runs before refactor:                                       
                                                                                
    $ go test -check.f=BenchmarkRegenerateCIDRPolicyRules -check.b              
    PASS: resolve_test.go:227: PolicyTestSuite.BenchmarkRegenerateCIDRPolicyRules        500           4157036 ns/op
    OK: 1 passed                                                                
    PASS                                                                        
    ok      github.com/cilium/cilium/pkg/policy     2.653s                      
                                                                                
Best result of five runs after refactor:                                        
                                                                                
    $ go test -check.f=BenchmarkRegenerateCIDRPolicyRules -check.b              
    PASS: resolve_test.go:227: PolicyTestSuite.BenchmarkRegenerateCIDRPolicyRules       1000           2361818 ns/op                                                                                                                                                                                                           
    OK: 1 passed                                                                
    PASS ok      github.com/cilium/cilium/pkg/policy     2.664s

```release-note
Improve policy import performance, particularly with CIDR policies
```