@joestringer 2022-01-11
[k8s-1.22-kernel-4.19 job](https://jenkins.cilium.io/job/Cilium-PR-K8s-1.22-kernel-4.19/248/) hit #18447, I've prepared a fix for it.
[k8s-1.23-kernel-net-next](https://jenkins.cilium.io/job/Cilium-PR-K8s-1.23-kernel-net-next/178/) hit #18386, will re-attempt that job.
[Travis job](https://app.travis-ci.com/github/cilium/cilium/jobs/555150463) had one successful run, and one run where it hit the pull ratelimit.